consumer/requirements.txt
kafka-python
kafka-python
psycopg2-binary